Introducing FusionPBX's auto phone provisioning that was written from scratch. The configuration files are template driven and easy to change. Templates are located in includes/provision_templates. More templates will be created and made available soon.
Voice -> Features -> Phones shows the list of hardware based phones. More can be added to the list as needed. This is where you would choose your template. After the choice has been made go to the extension and assign the phone from the select list to the extension and the line number. Initial provisioning is done over HTTP. other types of provisioning will be available in the future including ftp, and tftp.

git-svn-id: http://fusionpbx.googlecode.com/svn/trunk/fusionpbx@262 e2954ef0-caed-11de-bd98-e341d93cc3cc